+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **General Changes**
   - Setup initial flow through website
   - New users can now create accounts, and reset their passwords
+  - Account registration requires full name, email, date of birth, and recaptcha
   - New users are sent a welcome email that contains a link to create their
     password
   - Users can sign and accept consent form
@@ -11,6 +12,7 @@
   - An about page describes the study
   - A contact page was added to allow users to contact research study staff
   - Consent forms can now be downloaded as PDFs
+  - Users can reset their password on the settings page
   - Using Ruby on Rails 5.1.3.rc3 framework
 - **Admin Changes**
   - Admins can now manage user accounts